enhance-search: rename advanced_search_modal.tsx to AdvancedSearchModal.tsx
